feat: 完善连连支付 - 55%

+++ main/app/dto/resp/instant.go
@@ -251,9 +251,9 @@ type InstantHideSaleProductList struct {
 
 // 销售订单的支付页的二维码信息
 type InstantOrderPaymentQrcodeInfoResp struct {
-	PaymentOrderUuid uint64  `json:"payment_order_uuid"` // 支付单uuid
-	QrCode           string  `json:"qr_code"`            // 支付单二维码
-	QrCodeExpireSec  int     `json:"qr_code_expire_sec"` // 支付单二维码剩余时间（秒）(少于等于0的时候 需要重新生成二维码)
-	Status           int     `json:"status"`             // 支付单状态 支付状态, 0-未支付 1-已支付
+	PaymentOrderUuid uint64  `json:"payment_order_uuid"` // 支付单uuid (当/cashier/desk/order/payment/info接口的payment_orders中的存在相同的uuid时证明已经支付)
+	QrCode           string  `json:"qr_code"`            // 支付单二维码 (永远都是返回base64图片给前端直接显示)
+	QrCodeExpireSec  int64   `json:"qr_code_expire_sec"` // 支付单二维码剩余时间（秒）(少于等于0的时候 需要重新生成二维码, 再次请求当前接口就行)
+	Status           int     `json:"status"`             // 支付单状态 支付状态, 0-未支付 1-已支付 (可选择轮询当前接口，获取支付状态)
 	PaymentAmount    float64 `json:"payment_amount"`     // 支付金额
 }
